I moved all the interpreter code from Allegrex.isa to XXXState.java : - Allegrex.isa is smaller. - No need to alter Allegrex.java to implement an instruction. Now it must be in XXXState.java. Also I added a log error for each instruction, especially for VFPU so we can track which instruction is needed.

+++ src/jpcsp/Allegrex/BcuState.java
@@ -48,7 +48,7 @@ public class BcuState extends LsuState {
     }
     
     public static int branchTarget(int npc, int simm16) {
-        return npc + (((int)(short)simm16) << 2);
+        return npc + (simm16 << 2);
     }
 
     public static int jumpTarget(int npc, int uimm26) {
